фикс пагинатора, рендеринг страницы мувиАпи
