+++ src/index.html
@@ -23,25 +23,6 @@
           ></ul>
           <!-- Pagination Materialized -->
           <div class="pagination-container">
-            <ul class="pagination pagination_list">
-              <li class="waves-effect pagination-beginning">
-                <a href="#!"><i class="material-icons">first_page</i></a>
-              </li>
-              <li class="waves-effect pagination-previous-page">
-                <a href="#!"><i class="material-icons">chevron_left</i></a>
-              </li>
-              <li>
-                <ul class="pagination-page-items-container">
-                  <!-- Pagination page items are inserted here -->
-                </ul>
-              </li>
-              <li class="waves-effect pagination-next-page">
-                <a href="#!"><i class="material-icons">chevron_right</i></a>
-              </li>
-              <li class="waves-effect pagination-end">
-                <a href="#!"><i class="material-icons">last_page</i></a>
-              </li>
-            </ul>
           </div>
         </div>
       </section>
