+++ src/index.js
@@ -14,6 +14,8 @@ import './scss/main.scss';
 
 // Importing JS files
 import './js/movieApi.js';
+import './js/fetch-functions.js';
+import './js/paginator.js';
 import './js/show-details.js';
 import './js/buttonUp.js';
 import './js/navigation.js';
