+++ src/templates/8pagination.hbs
@@ -1,9 +1,9 @@
 <ul class="pagination pagination_list">
     <li class="waves-effect pagination-beginning {{ left }}">
-        <a href="#!"><i class="material-icons">first_page</i></a>
+        <a href="#!"><i class="material-icons" {{ left }}>first_page</i></a>
     </li>
     <li class="waves-effect pagination-previous-page {{ left }}">
-        <a href="#!"><i class="material-icons">chevron_left</i></a>
+        <a href="#!"><i class="material-icons" {{ left }}>chevron_left</i></a>
     </li>
     <li>
         <ul class="pagination-page-items-container">
@@ -15,9 +15,9 @@
         </ul>
     </li>
     <li class="waves-effect pagination-next-page {{ right }}">
-        <a href="#!"><i class="material-icons">chevron_right</i></a>
+        <a href="#!"><i class="material-icons" {{ right }}>chevron_right</i></a>
     </li>
     <li class="waves-effect pagination-end {{ right }}">
-        <a href="#!"><i class="material-icons">last_page</i></a>
+        <a href="#!"><i class="material-icons" {{ right }}>last_page</i></a>
     </li>
 </ul>
